refactor: move header files to their own folder
