+++ headers/Cliente.hpp
@@ -1,6 +1,5 @@
 #ifndef CLIENTE_H
 #define CLIENTE_H
-
 #include <string>
 
 using namespace std;
