+++ headers/EncomendaNormal.hpp
@@ -1,8 +1,8 @@
 #ifndef EncomendaNormal_H
 #define EncomendaNormal_H
 
-#include "../Encomenda.hpp"
-#include "../../cliente/Cliente.hpp"
+#include "Encomenda.hpp"
+#include "Cliente.hpp"
 
 using namespace std;
