+++ headers/EncomendaRelampago.hpp
@@ -1,8 +1,8 @@
 #ifndef EncomendaRelampago_H
 #define EncomendaRelampago_H
 
-#include "../Encomenda.hpp"
-#include "../../cliente/Cliente.hpp"
+#include "Encomenda.hpp"
+#include "Cliente.hpp"
 using namespace std;
 
 class EncomendaRelampago: public Encomenda {
