+++ src/EncomendaRelampago.cpp
@@ -1,11 +1,11 @@
-#include "../Encomenda.hpp"
-#include "EncomendaRelampago.hpp"
+#include "../headers/Encomenda.hpp"
+#include "../headers/EncomendaRelampago.hpp"
 #include <iostream>
 
 double EncomendaRelampago::calcula() {
   double result = this->getPeso() * this->getCustoKG();
-  result += result * 0.25;
-  return result;
+	result += result * 0.25;
+	return result;
 }
 
 void EncomendaRelampago::print(){
